+++ core/src/main/java/tuxkids/tuxblocks/core/defense/tower/HorizontalWall.java
@@ -16,7 +16,7 @@ public class HorizontalWall extends Tower {
 	}
 
 	@Override
-	public int damage() {
+	public float damage() {
 		return 0;
 	}
